renamed folders and namespace
